worked on the adminView page and adminOrders route,worked on checkout page,created Orders and ordersDetails page,orders schema

+++ src/app/api/admin/delete-product/route.ts
@@ -1,5 +1,7 @@
 import connectToDB from '@/database/Database';
+import AuthUser from '@/middleware/AuthUser';
 import Product from '@/modelsDB/Product';
+import { JwtPayload } from 'jsonwebtoken';
 import { NextRequest, NextResponse } from 'next/server';
 
 export const dynamic = 'force-dynamic';
@@ -8,37 +10,38 @@ export const DELETE = async (req: NextRequest) => {
   //Delete Data from DB
   try {
     await connectToDB();
-    // const isAuthUser = await AuthUser(req);
 
-    // if (isAuthUser?.role === "admin") {
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get('id'); //id of product to be deleted
+    const isAuthUser = (await AuthUser(req)) as JwtPayload;
 
-    if (!id)
-      return NextResponse.json({
-        success: false,
-        message: 'Product ID is required',
-      });
+    if (isAuthUser?.role === 'admin') {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get('id'); //id of product to be deleted
 
-    const deletedProduct = await Product.findByIdAndDelete(id); //to delete product by id from DB
+      if (!id)
+        return NextResponse.json({
+          success: false,
+          message: 'Product ID is required',
+        });
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: 'Product deleted successfully',
-      });
+      const deletedProduct = await Product.findByIdAndDelete(id); //to delete product by id from DB
+
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: 'Product deleted successfully',
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: 'Failed to delete the product ! Please try again',
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
-        message: 'Failed to delete the product ! Please try again',
+        message: 'You are not authenticated',
       });
     }
-    // } else {
-    //   return NextResponse.json({
-    //     success: false,
-    //     message: "You are not authenticated",
-    //   });
-    // }
   } catch (error) {
     console.log(error);
     return NextResponse.json({
